// src/app/data/interfaces/IUsuarios.ts
export interface IUsuarios {
  codUsuario?:          number;
  email?:               string;
  password?:            boolean;
  fechaCambioClave?:    boolean;
  estadoCuenta?:        Date;
  estadoClave?:         string;
  codPerfil?:           number;
  codPersona?:          number;
  codSucursal?:         number;
}